refactor: 린팅 규칙 적용 및 코드 정리
