+++ src/app/(auth-required)/main/page.tsx
@@ -1,8 +1,8 @@
 import { dehydrate, HydrationBoundary } from '@tanstack/react-query';
 import { Metadata } from 'next';
-import { PATHS } from '@/constants';
 import { postList, postSummary } from '@/apis';
-import { getQueryClient } from '@/utils/queryUtil';
+import { PATHS } from '@/constants';
+import { getQueryClient } from '@/utils';
 import { Content } from './Content';
 
 export const metadata: Metadata = {
@@ -22,17 +22,11 @@ export default async function Page({ searchParams }: IProp) {
   await client.prefetchInfiniteQuery({
     queryKey: [PATHS.POSTS, [searchParams.asc, searchParams.sort]],
     queryFn: async () =>
-      await postList({
-        asc: searchParams.asc === 'true',
-        sort: searchParams.sort || '',
-      }),
+      await postList({ asc: searchParams.asc === 'true', sort: searchParams.sort || '' }),
     initialPageParam: undefined,
   });
 
-  await client.prefetchQuery({
-    queryKey: [PATHS.SUMMARY],
-    queryFn: postSummary,
-  });
+  await client.prefetchQuery({ queryKey: [PATHS.SUMMARY], queryFn: postSummary });
 
   return (
     <HydrationBoundary state={dehydrate(client)}>
